Added documentation to Sample.py + removed all offline stuff
I do not entirely understand the offline component of the GLaSDI repository. I believe it's used when there is an external library available to solve the PDE. I think the idea is to use that rather than a hard-coded solve method (like in burgers1d.py).

Either way, the documentation is... non existent, so I don't really understand it. I don't entirey intend to use it either.

Thus, to remove clutter and code that I don't understand, I removed it. I left some of the code in but commented out. I deleted other parts. The code still appears to run, even though it has no offline stuff now.