Changed latent dynamics IC to a 2d array It has shape [n_IC, dim], where n_IC is the number of initial conditions we need to specify to pin down the initial state of the system (the number of derivatives we have to specify). dim is the dimension of the latent dynamics.
Kevin's code failed to account for the fact that we may want multiple IC's (if we have more than 1st order dynamics).

Making this change was... tricky. First I had to modify the latent dynamics files (including those in sindy) and the LD superclass.

This created a problem, however... GPLaSDI does NOT give IC's in the form [n_IC, dim]. It gives them in the form [dim]. I had to modify the initial_condition_latent function to fix this.

initial_condition_latent was tricky. I couldn't find a good way to modify it to work universally... so I made it a part of the model class. Each model now has to specify how to encode the latent space. This is fine. With this change, I made the autoencoder class map fom IC's to elements of shape [1, nz], where nz is the latent dimension. I then return a list of these kinds of arrays... this makes the autoencoder class play nicely with the new format. I then wrote a latent_initial_condition function for the Autoencoder_Pair function which returns a list of arrays of shape [2, nz] whose rows hold the initial position/velocity. This should make it play nicely with the new PairSpring function.

I have a lotttttt of code to test now. My plan is to try testing with a pair autoencoder rigged to work with burgers data. I will have it set the initial position as the IC for both autoencoders or something. This should allow me to test all of the code I wrote/begin debugging it. Expect bug fixes and such in the next few commits.

After that, the only thing I have left to write is a physics class... I jsut need to remember to have this return a tuple of ICs.