Changed X_Train and X_Test to be lists of tensors
I need this change to be able to run my position/velocity code for the 2nd order model. The issue is that making this change forced me to change a lottttt of code. This code base is very very very rigid.

I also generalized the GPLaSDI code to work on models other than Autoencoders... I think.

I still need to test everything.