Lots of bug fixes + AE/AE_Pair work in batches of params + IC is a list
I fixed a lot of bugs... a big one was using X_train and X_Train in different parts of the code to refer to the same variable. I fixed that.

I also changed the autoencoder/autoencoder_pair class to take in an input of the form (np, nt, ...), where, np is the number of parameter combinations we are testing, nt is the number of time steps in each solution sequence. This is a big departure from before, but it fixed a lot of issues.

I also changed the IC in the latent dynamics class to be a list, not a 2d array. This is better/fits with the format I have in the rest of the code (a lot of lists now). This change forced me to change a lot of code... god damn this code is rigid/difficult to modify. It's supposed to be generalizable... but that isn't the case. No one imagined using this code for anything other than 1st order dynamics. I have generalized it (mostly) to work with dynamics of arbitrary order.e

I still need to debug a lot, but things appear to be running, for the most part.