Lots of bug fixes Initial_Condition in Physics now returns a list of numpy ndarrays, not a single one. I generalized it to work with higher dimensional dynamics.
calibrate no longer takes a numpy argument. It always retuns a tensor for coefs... deal with it/change it yourself.

simulate now returns a list of arrays, again to deal with higher dimensional dynamics.

The code appears to run. I still need to debug it/make sure it is doing the right thing, but it can run without giving errors now.