Updated sample_roms and added new plotting functions I copied the plotting functions from postprocess.py in the original repo. I rewrote one of these functions and have two more to go... oops.
While rewritingt his function, I realized that sample_roms was designed to work assuming that the latent dynamics only had a single derivative/n_IC = 1. I changed this (so that it works for an arbitrary n_IC)... but doing so required a lot more changes than I expected. It should work now.

I have not debuged any of this code.