Huge overhaul of simulate function I want to write a rollout loss. To do that, I need to be able to simulate an entire batch of latent states forward a few time steps. In other words, when I encode a sequence of fom frames, I get a sequence of encoding frames. I want to treat each as an IC, then simulate it forward a few time steps (using the current dynamics), decode the updated latent frames, and then compare them to the corresponding fom frames. I could do this by making a loop over the latent frames/running simulate on each one... but that is very slow. A faster approach is to simulate the entire batch forward in time at once. The previous simulate function simply could not do that... so I massivly upgraded it.
Like calibrate, it can now work with a batch of parameter values, and a batch of frames for each parameter value. This should allow me to do everything I want to do. I believe the code works, though I have not debugged it/not revamped the rest of the code to use this new simulate function. Later, I plan to adapat the same kind of function to the sindy simulate function + make this new approach the default for the
latent dynamics class. But first, I need to get this thing running.