I fixed a bug in Physics
The t_Grid values need to be a torch.Tensor... not a numpy.ndarray... that changed since I started the rewrite (and Physics was the first thing I rewrote). I fixed this, and changed a few comments elsewhere... everything should be done now... I just need to rewrite the driver code (workflow, jupyter notebook) and debug everything.