Got added timers  + updated timing.
I added timers to GPLaSDI to try and figure out where all the runtime is going... turns out it's backprop and Rollout. I think most of the runtime in backprop is from rollout. I get to figure out how to fix this in the next commit, I guess.
The timer prints in a much nicer format now.