Updated nonlinear elastcity to use random points
This proved to be a bit more challenging because the mesh moves. I need to evaluate the GF's at the same material points... not the same physical point. I realized that you can do this fairly easily using element numbers and reference coordinates. Actually, the code with this approach is a lot cleaner. See my changes. I plan to use the same approach in the other MFEM examples. Expect that in the next commit