Rewrote Compute_Error_and_STD
SolveROMs: I rewrote Compute_Error_and_STD from scratch. The old function had a few bugs: If the true solution happened to be 0 (or very close to it) at some point, the relative error calculation blew up. Further, if a frame was all zeros (such as is the case with a lot of the initial velocities), the relavie error goes to infinity. I fixed this by rewriting the function to use a "vector relative error" (divide by the L^\inf norm of the current frame rather than a specific component of the current frame). I got this idea from William. It appears to have fixed my issue. I then added a checker that sets the divisor to 1 if we happen to get a frame of all zeros.

I made minor tweaks to the rest of the code to work with the new interface.