Bug fixes
I updated the plot titles for the SD / error heatmaps. I also found a bug in the function that computes the errors/STDs. In the "Autoencoder_Pair" case, it failed to populate the second column of the max_STD matrix. I fixed this. I also found settings that appear to make the Telegraphers example converge. I need to do the same for other examples... but I think everything is running now/giving reasonable results... which is the first time I have been able to say that in a while.

In a future commit, I want to fully document the non-linear elasticity example... though that may have to wait until after my Czechia trip.