Updated default variable values in Wave Equation
It turns out that the issue wasn't with the rest of the code... it was just that the wave equation evolved too quickly (and had too low a frame rate) for LaSDI to have any hope of learning it. I reduced the wave speed and bumped up the frame rate. This appears to have solved the issues I was having before; LaSDI converges well now.