Fixed bug in Error/Std function + added plot for relative error
There were several bugs in the relative error code. It wasn't normalized correctly... I fixed it. I also realized that the relative error in the velocity spikes in the first few timesteps. To fix this, I added a parameter that allows me to skip some proportion of the frames (to avoid this issue). The spike occurs because the initial veoocity is 0, meaning that the velocity in the first frame is tiny.. any prediction error leads to a massive relative velocity. This is misleading, so I decided to ignore the relative erorr in the first few frames.

I also added a plot for the relative error to the workflow.