Modified IC in PyMFEM scripts
The driver scripts now call the IC functions in the PyMFEM code directly. This eliminates the risk of having different IC functions. I also beefed up the IC functions in the PyMFEM scripts so that they can work with one set of coordinates (the way PyMFEM calls them) or a batch of them (the way the driver scripts call them).