Added a new "IC Rollout" loss
The idea is simple: For each training parameter, fetch its IC and encode it. Solve it forward for a certain number of time steps and then decode the entire latent trajectory. Compare this with the corresponding frames in the true IC. I also encode the true FOM frames and compare their encodings to the latent trajectory.

This creates a second kind of latent trajectory, one which should hopefully engender better generalization. We will see.