Fixed serialization + found a big bug
the IC functions do not set the parameters correctly. Something goes wrong with global. thus, when the IC function tries to load them, it just uses the last value it got from the simulate function or whatever. I fixed this by making the parameters (k or w) an explicit variable which can be passed to the IC objects. I think I want this to become the default way it is passed... no more global nonsense.

This fixed the issue. IDk why it didn't cause bigger issues before... I guess the parameter k just doesn't change much. I partially fixed the wave equation, I need to fix the rest later.

I also made serlization work now. Woo-hoo.... it was a bit of a mess.