Modified code to rescale training data
I compute the std of each time derivative of the solution for each combination of training parameters. When I compute the reconstruction, rollout, or IC_Rollout losses, I rescale the loss so that it is as if the target data has an STD of 1. This should help rescale the data so that it doesn't prioritize any derivatives or training combinations.