Changed relative error computation
The issue is that the old one would blow up if the demoinator approached zero at some frame... which happens. I think the new metric avoids a lot of the issues of the old one... we will see, I guess. I have no clue what a "good" relative error is now.