Rescaled time axis in nonlinear elasticity
The problem goes to t = 50. That is a huge range, and it causes some stability issues. The velocity was basically 0 before... which meant that my reascaling stuff was blowing up any velocity losses by several orders of magnitude. If the network made any errors, they got amplified. This was causing problems... big problems... the networks just couldn't learn the velocity well enough. The issue is that the velocity is just too small everywhere. I fixed it by rescaling the time axis to go from 0 to 5 (not 0 to 10). This required me to rescale the velcotiy by a factor of 10 and the time grid by a factor of .1. I think it has improved convergence.

The other changes are minor