Added a second explicit example
It now works in two dimensions.
I made a seperate "one time derivative" and "two time derivative" version of explicit.

The former now works in 2 spatial dimensions.

I also fixed a few bugs in workflow.

I am in the process of updating the assertions to report inofmration when they are triggered.