Changed i_random to i_worst in workflow
Previously, the code sampled a random point in the testing set, then made latent trajectory plots and animations and whatnot using that. I don't like that. I want it to be the worst one (the one that attains the highest rollout error). To do this, I changed the code to figure out which parameter combination yields the highest rollout error (across all IC terms), then makes animations + latent plots for that one. In other words, I replaced i_random with a new index, i_worst.