Added plot to test for differences between training parameters
This should let me detect if the range I selected for a paramter is too big or not (if the relative error is huge when we change one parameter, we probably need to make the range tightr). Changing 's' from 0.9 to 1.1 in nonlinear elasticity induces a relaive error (100 * ||U_i - U_j||_2 / ||U_j||_2) by like 100%!!!!! I need to make that range tighter. Anyway, now I know why this thing has been struggling.