Added "Plot training set relative error heatmap" option
If true, we plot the relative error heatmap after each cycle of greedy sampling; this is useful for debugging purposes. I also added code to make the heatmap with the final training set when the code makes all of the plots. I had to update all yml files to include the new setting.